Basic AWS CLI Compatibility
Being able to create a bucket, add items to it, and manage the deletion of both it and the bucket is basic aws cli functionality.

To diagnose operations as a user, the listing of buckets and their keys is supported as well.

This does not support any extension of the basic behavior of any of these features (ex: ACLs)